requirements.txt
flask
flask-cors
scikit-learn
pandas
numpy
joblib
imbalanced-learn
matplotlib
PyJWT
python-dotenv
requests
flask-sqlalchemy
pytest
celery
redis
twilio
